+++ app/src/androidTest/java/grp3/projet/miage/paris10/fr/m1c20152016_planinteractifbu/ApplicationTest.java
@@ -1,4 +1,4 @@
-package m1c20152016_planinteractifbu.fr.paris10.miage.projet.grp3.planbu;
+package grp3.projet.miage.paris10.fr.m1c20152016_planinteractifbu;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
