+++ app/src/main/AndroidManifest.xml
@@ -4,13 +4,13 @@
 
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/books"
-        android:label="Plan interactif"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="m1c20152016_planinteractifbu.fr.paris10.miage.projet.grp3.planbu.MainActivity"
-            android:label="Plan interactif"
+            android:name=".MainActivity"
+            android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
