+++ app/src/main/res/layout/content_main.xml
@@ -9,8 +9,8 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="m1c20152016_planinteractifbu.fr.paris10.miage.projet.grp3.planbu.MainActivity"
-    tools:showIn="@layout/app_bar_main">
+    tools:context="grp3.projet.miage.paris10.fr.m1c20152016_planinteractifbu.MainActivity"
+    tools:showIn="@layout/activity_main">
 
     <TextView
         android:layout_width="wrap_content"
