+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,4 @@
 <resources>
     <string name="app_name">m1c20152016-planinteractifbu</string>
     <string name="action_settings">Settings</string>
-
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
 </resources>
