--- app/src/test/java/m1c20152016_planinteractifbu/fr/paris10/miage/projet/grp3/planbu/ExampleUnitTest.java
@@ -1,15 +1,0 @@
-package m1c20152016_planinteractifbu.fr.paris10.miage.projet.grp3.planbu;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
