+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':app', ':app', ':app'
+include ':app'
